+++ src/tests/integration.test.ts
@@ -15,11 +15,11 @@ describe('Direct usage', () => {
 });
 
 describe('Root Level', () => {
-    it('should basicly work for :root - pseudo selector', async () => {
+    it('should basically work for :root - pseudo selector', async () => {
         await testFixture('root_level/root');
     });
 
-    it('should basicly work for body selector', async () => {
+    it('should basically work for body selector', async () => {
         await testFixture('root_level/body');
     });
 });
@@ -29,7 +29,7 @@ describe('Permutation', () => {
         await testFixture('permutation/class');
     });
 
-    it("should use global variables if the varaible isn't avaiable", async () => {
+    it("should use global variables if the variable isn't available", async () => {
         await testFixture('permutation/multiple_variables');
     });
 });
@@ -37,5 +37,17 @@ describe('Permutation', () => {
 describe('Media Queries', () => {
     it('should support media queries', async () => {
         await testFixture('media_queries/simple');
+    });
+
+    it('should support simple multiple media queries', async () => {
+        await testFixture('media_queries/multiple_media_queries');
+    });
+
+    it('should support media query with multiple variables', async () => {
+        await testFixture('media_queries/media_query_multiple_variables');
+    });
+
+    it('should support multiple (complex) media queries', async () => {
+        await testFixture('media_queries/multiple_complex');
     });
 });